+++ src/main/java/com/co/web/avanzada/repository/IProductoRepo.java
@@ -11,7 +11,7 @@ import com.co.web.avanzada.entity.Producto;
 
 @Repository
 public interface IProductoRepo extends
-JpaRepository<Producto, Long>{
+JpaRepository<Producto, Integer>{
 	@Query("Select P from Producto P, Inventario I where I.producto.codigoProducto= P.codigoProducto and P.categoria.idCategoria=?1")
 	List<Producto> findByCategoria(int idCategoria);
